Automatic control of the delay line
First, made a copy of the functions.py to a file named old-functions.py, which I can still use if I have problems with the new file (I didn't want to branch here, because I need it as a legacy option, for now).

As for the modifications is the functions.py, I added the KBD101 class to control the delay line via Python code, and modified the Measurement class for it to control the delay line by itself.