Implemented zero position and custom colormap
Now, when plotting the time domain data, it's possible to define the position in the delay line we want to set the time zero to. It is particularly important when comparing data with different delayline ranges.

Also, for this to be accomplished, the Data instances have now and idn attribute which is an Identity object.

Last, it's now possible to choose custom cm colormaps when plotting with Plot.